Fixed the checksum calculations. There were issues with casting.

+++ icmp.c
@@ -62,7 +62,7 @@ void icmp_response(struct packet_state *ps, struct ip *ip_hdr, unsigned int type
 		case ICMPT_ECHOREPLY:
 			res_head = create_icmp_hdr(ps, type, code);
 			len = ps->res_len - sizeof(struct sr_ethernet_hdr) - sizeof(struct ip);
-			res_head->icmp_sum = cksum((uint8_t *) res_head, len);
+			res_head->icmp_sum = cksum((uint16_t *) res_head, len);
 			
 			break;
 
@@ -70,14 +70,14 @@ void icmp_response(struct packet_state *ps, struct ip *ip_hdr, unsigned int type
 			res_head = create_icmp_hdr(ps, type, code);
 			create_icmp_data(ps, ip_hdr);
 			len = ps->res_len - sizeof(struct sr_ethernet_hdr) - sizeof(struct ip);
-			res_head->icmp_sum = cksum((uint8_t *) res_head, len);
+			res_head->icmp_sum = cksum((uint16_t *) res_head, len);
 			break;
 
 		case ICMPT_TIMEEX:
 			res_head = create_icmp_hdr(ps, type, code);
 			create_icmp_data(ps, ip_hdr);
 			len = ps->res_len - sizeof(struct sr_ethernet_hdr) - sizeof(struct ip);
-			res_head->icmp_sum = cksum((uint8_t *) res_head, len);
+			res_head->icmp_sum = cksum((uint16_t *) res_head, len);
 			break;
 
 		case ICMPT_TRACERT:
